+++ resources/js/app.js
@@ -1,10 +1,6 @@
 import './bootstrap';
-
-import Alpine from 'alpinejs';
 import focus from '@alpinejs/focus';
 
-window.Alpine = Alpine;
-
-Alpine.plugin(focus);
-
-Alpine.start();
+document.addEventListener('alpine:init', () => {
+    window.Alpine.plugin(focus);
+});